+++ src/lib/Sema/TensorType.cpp
@@ -17,7 +17,7 @@ bool TensorType::operator==(const TensorType &rhs) const {
 const std::string TensorType::getDimString() const {
   std::string result = "";
   for (unsigned i = 0; i < getRank(); i++) {
-    result += std::to_string(getDim(i));
+    result += std::to_string((long long)getDim(i));
     if (i != (getRank()-1))
       result += ", ";
   }
